refactor TFLite backend to support multiple ML pipelines
