+++ mobile_back_tflite/cpp/backend_tflite/BUILD
@@ -38,8 +38,15 @@ pbtxt2header(
 
 cc_library(
     name = "tflite_c",
-    srcs = ["tflite_c.cc"],
+    srcs = [
+        "single_model_pipeline.cc",
+        "stable_diffusion_pipeline.cc",
+        "tflite_c.cc",
+    ],
     hdrs = [
+        "pipeline.h",
+        "single_model_pipeline.h",
+        "stable_diffusion_pipeline.h",
         "tflite_settings_android.h",
         "tflite_settings_apple.h",
         "tflite_settings_windows.h",
